scripts/requirements.txt
absl-py==1.0.0
addict==2.4.0
aiohttp==3.8.1
aiosignal==1.2.0
api==0.0.7
argon2-cffi==20.1.0
astropy==5.0.1
astunparse==1.6.3
async-generator==1.10
async-timeout==4.0.2
asyncio==3.4.3
attrs==21.2.0
backcall==0.2.0
backports.entry-points-selectable==1.1.0
basicsr==1.3.4.9
bcc==0.18.0
beautifulsoup4==4.9.3
black @ git+https://github.com/psf/black.git@8ea641eed5b9540287a8e9a9afa1458b72b9b630
bleach==4.0.0
blinker==1.4
blis==0.7.4
blobfile==1.2.8
Brlapi==0.8.2
bs4==0.0.1
cachetools==4.2.4
casttube==0.2.1
catalogue==2.0.4
certifi==2021.10.8
cffi==1.14.6
chardet==3.0.4
charset-normalizer==2.0.11
chex==0.1.3
chrome-gnome-shell==0.0.0
chromedriver-installer==0.0.6
click==7.1.2
-e git+https://github.com/openai/CLIP@40f5484c1c74edd83cb9cf687c6ab92b28d8b656#egg=clip
colorama==0.4.4
command-not-found==0.3
cryptography==3.4.7
cupshelpers==1.0
cycler==0.10.0
cymem==2.0.5
dalle-mini==0.1.0
DateTime==4.3
dbus-python==1.2.16
debugpy==1.4.1
decorator==5.0.9
defer==1.0.6
defusedxml==0.7.1
distlib==0.3.2
distro==1.6.0
dm-tree==0.1.7
dnspython==2.1.0
docker-pycreds==0.4.0
ecdsa==0.17.0
einops==0.4.1
emoji==1.7.0
entrypoints==0.3
evdev==1.4.0
facexlib==0.2.0.3
fastai==2.5.3
fastbook==0.0.18
fastcore==1.3.27
fastdownload==0.0.5
fastprogress==1.0.0
fastrelease==0.1.12
filelock==3.0.12
filterpy==1.4.5
Flask==2.0.1
flatbuffers==1.12
flax==0.5.0
frozenlist==1.2.0
ftfy==6.1.1
fudge==1.1.1
future==0.18.2
gast==0.4.0
geographiclib==1.52
geopy==2.2.0
gevent==21.12.0
gfpgan==0.2.1
ghapi==0.1.19
gitdb==4.0.9
GitPython==3.1.27
google-auth==1.35.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
gpg===1.14.0-unknown
graphviz==0.19.1
greenlet==1.1.0
grequests==0.6.0
grpcio==1.43.0
-e git+https://github.com/crowsonkb/guided-diffusion@7a1c0dd8f17e5d59814e1515b86d2bf8baf3e840#egg=guided_diffusion
h5py==3.1.0
hidpidaemon==18.4.6
html5lib==1.1
httplib2==0.19.1
huggingface-hub==0.7.0
idna==2.8
ifaddr==0.1.7
imageio==2.14.1
importlib-metadata==4.10.1
iniconfig==1.1.1
ipykernel==6.0.3
ipython==7.26.0
ipython-genutils==0.2.0
ipywidgets==7.6.3
itsdangerous==2.0.1
jax==0.3.13
jaxlib==0.3.10+cuda11.cudnn82
jedi==0.18.0
jeepney==0.7.1
Jinja2==3.0.1
jmd-imagescraper==1.0.2
joblib==1.0.1
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.12
jupyter-console==6.4.0
jupyter-core==4.7.1
jupyter-http-over-ws==0.0.8
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.0
keras==2.7.0
Keras-Applications==1.0.8
keras-nightly==2.5.0.dev2021032900
Keras-Preprocessing==1.1.2
keras-vggface==0.6
kernelstub==3.1.4
keyring==23.0.1
kiwisolver==1.3.1
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.3
lazr.uri==1.0.5
ldap3==2.9.1
llvmlite==0.37.0
lmdb==1.3.0
louis==3.18.0
lpips==0.1.4
lxml==4.6.3
lz4==4.0.0
macaroonbakery==1.3.1
Mako==1.1.4
Markdown==3.3.6
MarkupSafe==2.0.1
matplotlib==3.4.2
matplotlib-inline==0.1.2
mediapipe==0.8.9.1
mistune==0.8.4
mkchromecast==0.3.9
Mopidy==3.2.0
Mopidy-SoundCloud==3.0.1
more-itertools==8.8.0
MouseInfo==0.1.3
msgpack==1.0.2
multidict==5.2.0
murmurhash==1.0.5
mypy-extensions==0.4.3
mysqlclient==1.4.4
natsort==7.1.0
nbclient==0.5.3
nbconvert==5.6.1
nbdev==1.1.19
nbformat==5.1.3
neovim==0.3.1
nest-asyncio==1.5.1
netaddr==0.8.0
netifaces==0.11.0
networkx==2.6.3
nose==1.3.7
notebook==6.4.0
numba==0.54.0
numpy==1.22.2
oauthlib==3.2.0
olefile==0.46
onboard==1.4.1
onnx==1.11.0
opencv-contrib-python==4.5.5.62
opencv-python==4.5.5.62
opt-einsum==3.3.0
optax==0.1.2
packaging==21.3
pandas==1.3.1
pandocfilters==1.4.3
parso==0.8.2
pathlib==1.0.1
pathspec==0.9.0
pathtools==0.1.2
pathy==0.6.0
pbr==5.6.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.0.1
platformdirs==2.2.0
pluggy==0.13.1
pop-transition==1.1.2
portpicker==1.2.0
preshed==3.0.5
prettyprint==0.1.5
prometheus-client==0.11.0
promise==2.3
prompt-toolkit==3.0.27
protobuf==3.19.4
psutil==5.8.0
ptyprocess==0.7.0
py==1.10.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyAutoGUI==0.9.53
pycairo==1.20.1
PyChromecast==9.2.0
pycparser==2.20
pycrypto==2.6.1
pycryptodomex==3.10.1
pycups==2.0.1
pydantic==1.8.2
pydbus==0.6.0
pyerfa==2.0.0.1
PyGetWindow==0.0.9
Pygments==2.9.0
PyGObject==3.40.1
pyheif==0.6.1
pyinotify==0.9.6
PyJWT==2.1.0
pykka==3.0.1
pymacaroons==0.13.0
PyMsgBox==1.0.9
PyNaCl==1.4.0
pynvim==0.4.3
pyOpenSSL==20.0.1
pyparsing==3.0.7
pyperclip==1.8.2
PyQt5==5.15.4
PyQt5-sip==12.9.0
PyRect==0.2.0
pyRFC3339==1.1
pyrsistent==0.18.0
PyScreeze==0.1.28
PySide2==5.15.1
pytesseract==0.3.9
pytest==6.2.4
python-apt===2.2.1pop0-1628521831-21.10-5dbe6cf
python-dateutil==2.8.2
python-debian==0.1.40
python-magic==0.4.24
python-xlib==0.31
python3-xlib==0.15
pytweening==1.0.4
pytz==2021.1
PyWavelets==1.2.0
pyxattr==0.7.2
pyxdg==0.27
PyYAML==6.0
pyzmq==22.2.0
qtconsole==5.1.1
QtPy==1.9.0
realesrgan==0.2.2.3
regex==2021.8.3
repolib==1.5.2
repoman==1.4.0
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.8
scikit-image==0.19.1
scikit-learn==0.24.2
scipy==1.8.0
screen-resolution-extra==0.0.0
seaborn==0.11.2
SecretStorage==3.3.1
selenium==3.141.0
Send2Trash==1.7.1
sentencepiece==0.1.96
sentry-sdk==1.5.10
sessioninstaller==0.0.0
setproctitle==1.2.2
shiboken2==5.15.1
shortuuid==1.0.8
simplegeneric==0.8.1
simplejson==3.17.3
six==1.16.0
sklearn==0.0
smart-open==5.1.0
smmap==5.0.0
snakeviz==2.1.1
soundcloud==0.5.0
soundcloud-cli==0.2.7
soupsieve==2.2.1
sourcery-cli==0.11.5
spacy==3.1.1
spacy-legacy==3.0.8
speedtest==0.0.1
SQLAlchemy==1.4.22
sqlalchemy-migrate==0.13.0
sqlparse==0.4.1
srsly==2.4.1
ssh-import-id==5.11
sty==1.0.0rc2
system76driver==20.4.56
systemd-python==234
tb-nightly==2.9.0a20220203
Tempita==0.5.2
tensorboard==2.5.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorboardX==2.4
tensorflow==2.5.0
tensorflow-estimator==2.5.0
termcolor==1.1.0
terminado==0.10.1
tesseract==0.1.3
testpath==0.5.0
thinc==8.0.8
thop==0.0.31.post2005241907
threadpoolctl==2.2.0
tifffile==2022.2.2
timm==0.5.4
tokenizers==0.12.1
toml==0.10.2
tomli==1.2.0
toolz==0.11.2
torch==1.9.1
torchvision==0.10.1
tornado==6.1
tqdm==4.62.3
traitlets==5.0.5
transformers==4.19.3
typer==0.3.2
typing_extensions==4.2.0
ubuntu-advantage-tools==27.8
ubuntu-drivers-common==0.0.0
ufw==0.36.1
Unidecode==1.3.4
urllib3==1.26.8
virtualenv==20.7.0
vqgan-jax @ git+https://github.com/patil-suraj/vqgan-jax.git@1be20eee476e5d35c30e4ec3ed12222018af8ce4
wadllib==1.3.5
wandb==0.12.15
wasabi==0.8.2
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==2.0.2
widgetsnbextension==3.5.1
wrapt==1.12.1
wxPython==4.0.7
xdg==5.1.1
xkit==0.0.0
xmltodict==0.12.0
yapf==0.32.0
yarl==1.7.2
youtube-dl==2021.6.6
youtube-downloader==0.6.2
zeroconf==0.33.2
zipp==3.7.0
zope.event==4.5.0
zope.interface==5.4.0
